materials: only export generic material attributes (fixes texture paths)
this is a potentially breaking change as it changes the set of generated material primitive attributes.
on the upside, it fixes some cases where the texture paths were exported as the CGA keys, not the final filesystem paths.

see prtx/Material.h for details regarding "CGA Material Keys" vs "PRTX Shader Material Keys"