Use https url to avoid need for SSH credentials.
Since this is a public repository and since the
build only reads from the repo, it makes sense
to avoid the need to manage credentials for the
repository and use the https url for cloning.

If at a later point we decide that we need write
access to the repo we can switch to using a deploy
key.